refactor: learner bulk import rename
