+++ apps/spaces/src/components/QuizViewLayout/index.tsx
@@ -25,6 +25,7 @@ import { DeleteModal } from "../modals/DeleteModal";
 import toast from "react-hot-toast";
 import { useQuestionCRUD } from "../../fetch/question";
 import { UnpublishedQuizModal } from "../modals/UnpublishedQuizModal";
+import { UnpublishQuizWithQuestionsModal } from "../modals/UnpublishQuizWithQuestionsModal";
 import { handleCopyUrlAndNotify } from "../../utils/quiz";
 import { getQuizResults, PublicQuizResultsResponse } from "../../fetch/results";
 import { useTranslation } from "react-i18next";
@@ -71,6 +72,7 @@ export const QuizViewLayout: FunctionComponent<Props> = () => {
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const [isRenameModalOpen, setIsRenameModalOpen] = useState(false);
   const [isUnpublishedQuizModalOpen, setIsUnpublishedQuizModalOpen] = useState(false);
+  const [isUnpublishQuizModalOpen, setIsUnpublishQuizModalOpen] = useState(false);
   const [showPublishTooltip, setShowPublishTooltip] = useState(false);
 
   const { destroy } = useQuestionCRUD()
@@ -228,7 +230,11 @@ export const QuizViewLayout: FunctionComponent<Props> = () => {
                       isEnabled={isPublished}
                       onToggle={() => {
                         if (disablePublishToggle) { return }
-                        handleTogglePublished(quiz.id, !quiz.published)
+                        if (isPublished && hasQuestions) {
+                          setIsUnpublishQuizModalOpen(true);
+                          return;
+                        }
+                        handleTogglePublished(quiz.id, !isPublished)
                       }}
                       rightLabel={t('quiz.publish_toggle.published')}
                       leftLabel={t('quiz.publish_toggle.unpublished')}
@@ -357,6 +363,14 @@ export const QuizViewLayout: FunctionComponent<Props> = () => {
                 }}
                 onCancel={() => {
                   handleCopyUrlAndNotify(quiz.hash, t('success_messages.quiz_link_copied'));
+                }}
+              />
+
+              <UnpublishQuizWithQuestionsModal
+                isModalOpen={isUnpublishQuizModalOpen}
+                setIsModalOpen={setIsUnpublishQuizModalOpen}
+                onConfirm={() => {
+                  handleTogglePublished(quiz.id, false);
                 }}
               />
